msm: kgsl: Use DMA APIs for memory pool cache maintenance
After allocating and zeroing pages from system memory or the pool we need
to ensure that the cache is synchronized so that it doesn't cause problems
down the road.  Use dma_sync_sg_for_device to make sure the allocated pages
are clean. This isn't the best way to handle this but we haven't yet come
up with a better way and this does the job.

Change-Id: Ic0dedbade48b700015bec172cf9b64e436364b4a
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
[Reina: Adapt for 4.19]
Signed-off-by: //. <174224977+skye-tachyon@users.noreply.github.com>